Refactored Inning to be a list of event states
