+++ src/models/Inning.ts
@@ -1,0 +1,5 @@
+import AtBatState from "./AtBatState";
+
+export default interface Inning {
+  events: AtBatState[];
+}
